Design Model for Publishing an E-Book
This commit includes the application service for the use case "Publish
comment as E-Book". The model supports just that use case. Publishing
means in this case uploading the generated E-Book. There's also an EBookPublished
event emmitted. This event can be used by chat bot to update the thread.

I deliberately kept the model simple. I.e. it doesn*t support any other
use cases (e.g. including multiple comments in one E-Book). I did this
to ensure that a simple MVP can be finished. If the MVP proves to be
successful, the model can then be redesigned to support other use cases
as well.

And one word to EBookFactory.generateUnpublishedFormats(). I chose this
design because of the learnings from the PoC. I can easily generate the
EPUB format, as it is a public format. However Kindle's format isn't. To
generate a Kindle format I need the EPUB format as input (which can then
be used with the proprietary kindlegen binary). That's why I designed
this particular method to take care of all formats with one call. As it
has the responsibility to create all formats, its implementation can use
dependencies between the generated formats (Kindle depends on EPUB).