Design model for asking a WP author for permission
This commit includes a model and application service which finds new
top comments of eligible writing prompts and post an answer to that
comment in which it asks for their permission to create the E-Book.

Although it seems a natural fit to model the dialog and its process as a
state machine (or process in BPMN) I do go for an event driven approach.
This allows me to design the bot stateless. A stateless bot has the
advantage to run without any persistence. This allows easy restarts (for
e.g. patching) and lowers the system requirements.