Design chat bot
Ok this is a bit much for one commit. This commit supports the two major
use cases "Ask for permission" and "Start publication on author's
permission". The model is quiet huge so chances are high that it might
change in the future. I hope so. It's a bit flexible so that more
chat commands can be supported in the future (Unpublish is on the list).