Add minimum age constraint to asking new WPs
This adds also a time window in which
AskPermissionService.askNewTopCommentsAuthorForPermission() operates.
Previously it would just remember an instant since the last iteration.
This would lead to skipping potential new WPs as they will be discarded for
their little score. At the time when their score would be enough they
are no more part of any batch. With the introduction of a minimum age,
askNewTopCommentsAuthorForPermission() shifts a time window which starts
at the end of the last window until now minus the minimum age.